fix bug

+++ src/main/java/org/stylehub/backend/e_commerce/modules/dashboard/brand_owner/catalog/dto/product/FindAllProductForBrand.java
@@ -11,6 +11,6 @@ public record FindAllProductForBrand(
         String categoryNameEn,
         String categoryNameAr,
         Long currentColors,
-        Integer basePrice
+        BigDecimal basePrice
 ) {
 }
